+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/option/SoraAudioOption.kt
@@ -16,12 +16,16 @@ class SoraAudioOption {
         const val HIGH_PASS_FILTER_CONSTRAINT = "googHighpassFilter"
         const val NOISE_SUPPRESSION_CONSTRAINT = "googNoiseSuppression"
     }
+
+    // TODO(zztkm): 破壊的変更にはなるが、DEFAULT を先頭に持ってくる
     /**
      * 利用できる音声コーデックを示します.
      */
     enum class Codec {
         /** Opus */
         OPUS,
+        /** Sora のデフォルト値を利用 */
+        DEFAULT,
     }
 
     /**