+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/error/SoraDisconnectReason.kt
@@ -7,6 +7,10 @@ enum class SoraDisconnectReason(val value: String?) {
     NO_ERROR("NO-ERROR"),
     WEBSOCKET_ONCLOSE("WEBSOCKET-ONCLOSE"),
     WEBSOCKET_ONERROR("WEBSOCKET-ONERROR"),
+    /**
+     * DataChannel シグナリングのみ利用時に Sora から切断が発生した
+     */
+    DATACHANNEL_ONCLOSE("DATACHANNEL-ONCLOSE"),
     PEER_CONNECTION_STATE_FAILED(null),
     SIGNALING_FAILURE(null),
 }